Add deprecation notice
Over time, zephyr has incorporated many of my original ideas from
zsh-utils while adopting a similar methodology - stick to configuring
features already available on your system (both in zsh and configurable
via environment variables) and making them easier to use while avoiding
unnecessary options.

At this point, zephyr is better in almost every way - it has better
defaults, more modules, more workarounds for oddities in zsh,
and is better maintained in general.

It's a little odd sunsetting this framework because I originally saw
this as the "final" version of my personal config, but it's time. I've
been working on zsh config frameworks in some form for almost 10 years,
starting with prezto and moving on to zsh-utils and I'm ready to let it
go.

Part of that is a testament to the simplicy in design, the stability
that can lead to, and the limited need for change when code works well.

I'm still very proud of what I built, and I'm glad the ideas live on,
but there are better options out there, and as I've had limited time to
continue development, I'm adding this deprecation notice.

For any PRs which are submitted for tweaks or features, I'm happy to
review and merge them, but support will be limited.

Major thanks to @mattmc3 for all his help in the past and his work on
zephyr.